Fix up problems in declarations from tsc
