+++ package.json
@@ -103,7 +103,7 @@
     "build:es6": "rollup -c --environment target:es6",
     "build:profiler": "rollup -c --environment target:profiler",
     "build:treemap": "npm run build:es5 -- --environment treemap",
-    "build:types": "tsc && rollup -c --environment target:types",
+    "build:types": "tsc && node types.mjs && rollup -c --environment target:types",
     "watch:debug": "npm run build:debug -- -w",
     "watch:es5": "npm run build:es5 -- -w",
     "watch:es6": "npm run build:es6 -- -w",
